+++ geotrek/common/urls.py
@@ -1,9 +1,7 @@
 from django.urls import path, converters, register_converter
 from mapentity.registry import MapEntityOptions
 
-from .views import (JSSettings, DocumentPublic, DocumentBookletPublic, import_view,
-                    import_update_json, ThemeViewSet, MarkupPublic, sync_view, sync_update_json, SyncRandoRedirect,
-                    CheckExtentsView)
+from . import views
 
 
 class LangConverter(converters.StringConverter):
@@ -14,21 +12,21 @@ register_converter(LangConverter, 'lang')
 
 app_name = 'common'
 urlpatterns = [
-    path('api/settings.json', JSSettings.as_view(), name='settings_json'),
-    path('tools/extents/', CheckExtentsView.as_view(), name='check_extents'),
+    path('api/settings.json', views.JSSettings.as_view(), name='settings_json'),
+    path('tools/extents/', views.CheckExtentsView.as_view(), name='check_extents'),
     path('commands/import-update.json', import_update_json, name='import_update_json'),
-    path('commands/import', import_view, name='import_dataset'),
-    path('commands/sync', SyncRandoRedirect.as_view(), name='sync_randos'),
-    path('commands/syncview', sync_view, name='sync_randos_view'),
-    path('commands/statesync/', sync_update_json, name='sync_randos_state'),
-    path('api/<lang:lang>/themes.json', ThemeViewSet.as_view({'get': 'list'}), name="themes_json"),
+    path('commands/import', views.import_view, name='import_dataset'),
+    path('commands/sync', views.SyncRandoRedirect.as_view(), name='sync_randos'),
+    path('commands/syncview', views.SyncRandoFormView.as_view(), name='sync_randos_view'),
+    path('commands/statesync/', views.sync_update_json, name='sync_randos_state'),
+    path('api/<lang:lang>/themes.json', views.ThemeViewSet.as_view({'get': 'list'}), name="themes_json"),
 ]
 
 
 class PublishableEntityOptions(MapEntityOptions):
-    document_public_view = DocumentPublic
-    document_public_booklet_view = DocumentBookletPublic
-    markup_public_view = MarkupPublic
+    document_public_view = views.DocumentPublic
+    document_public_booklet_view = views.DocumentBookletPublic
+    markup_public_view = views.MarkupPublic
 
     def scan_views(self, *args, **kwargs):
         """ Adds the URLs of all views provided by ``PublishableMixin`` models.